+++ src/filters.js
@@ -1,8 +1,6 @@
 // SVG Filter
 
 module.exports = function(p5) {
-    var SVGFilters = require('./p5.SVGFilters')(p5);
-
     var _filter = p5.prototype.filter;
     p5.prototype.filter = function(operation, value) {
         var svg = this._graphics.svg;